Refactor source context generation in implicits
